feat: apply box-sizing style to every Mokcups. WIP: add  frameOnly props
