+++ develop/src/demo/IosDemo.tsx
@@ -3,8 +3,8 @@ import { useCallback, useMemo, useState } from "react";
 import ButtonGroup from "../components/ButtonGroup";
 import ColorButton from "../components/ColorButton";
 import InputButton from "../components/InputButton";
-// import { IPadMockup, IPhoneMockup } from "../dist";
-import { IPadMockup, IPhoneMockup } from "react-device-mockup";
+import { IPadMockup, IPhoneMockup } from "../dist";
+// import { IPadMockup, IPhoneMockup } from "react-device-mockup";
 import CodeBlock from "./CodeBlock";
 import ScreenDemo from "./ScreenDemo";
 import demoStyle from "./demo.module.css";
