+++ src/ios-mockup/variants/phone/IPhoneIslandLandscape.tsx
@@ -28,6 +28,7 @@ const getStyles = (
 	return StyleSheet.create({
 		container: {
 			display: "flex",
+			boxSizing: "content-box",
 			position: "relative",
 			width: widthAndFrame,
 			height: heightAndFrame,
@@ -151,6 +152,7 @@ const getStyles = (
 
 export default function IPhoneIslandLandscape(props: PropsWithChildren<IIosMockupVariantProps>) {
 	const {
+		screenWidth,
 		frameColor,
 		statusbarColor,
 		hideStatusBar,
@@ -158,8 +160,8 @@ export default function IPhoneIslandLandscape(props: PropsWithChildren<IIosMocku
 		transparentNavigationBar,
 	} = props;
 	const styles = useMemo(
-		() => getStyles(props.screenWidth, frameColor, statusbarColor),
-		[props.screenWidth, frameColor, statusbarColor],
+		() => getStyles(screenWidth, frameColor, statusbarColor),
+		[screenWidth, frameColor, statusbarColor],
 	);
 
 	return (
