+++ src/ios-mockup/variants/phone/IPhoneLegacyPortrait.tsx
@@ -29,6 +29,7 @@ const getStyles = (
 		container: {
 			display: "flex",
 			flexDirection: "column",
+			boxSizing: "content-box",
 			position: "relative",
 			width: widthAndFrame,
 			height: mHeight + upperBezelHeight + lowerBezelHeight,
