+++ src/ios-mockup/variants/tab/IPadModernPortrait.tsx
@@ -29,6 +29,7 @@ const getStyles = (
 		container: {
 			display: "flex",
 			flexDirection: "column",
+			boxSizing: "content-box",
 			position: "relative",
 			width: widthAndFrame,
 			height: heightAndFrame,
@@ -130,6 +131,7 @@ const getStyles = (
 
 export default function IPadModernPortrait(props: PropsWithChildren<IIosMockupVariantProps>) {
 	const {
+		screenWidth,
 		frameColor,
 		statusbarColor,
 		hideStatusBar,
@@ -137,8 +139,8 @@ export default function IPadModernPortrait(props: PropsWithChildren<IIosMockupVa
 		transparentNavigationBar,
 	} = props;
 	const styles = useMemo(
-		() => getStyles(props.screenWidth, frameColor, statusbarColor),
-		[props.screenWidth, frameColor, statusbarColor],
+		() => getStyles(screenWidth, frameColor, statusbarColor),
+		[screenWidth, frameColor, statusbarColor],
 	);
 
 	return (
